Refactor split allocation Filters to component
